+++ src/Lykke.RabbitMqBroker/ArgumentsBuilderExtensions.cs
@@ -6,13 +6,13 @@ namespace Lykke.RabbitMqBroker;
 
 internal static class ArgumentsBuilderExtensions
 {
-    public static uint ToExpirationMilliseconds(this double src) => src switch
+    public static ulong ToExpirationMilliseconds(this double src) => src switch
     {
-        > uint.MaxValue => throw new ArgumentOutOfRangeException(nameof(src), "Value is too large to be converted to uint"),
-        < uint.MinValue => throw new ArgumentOutOfRangeException(nameof(src), "Value is too small to be converted to uint"),
-        _ => (uint)Math.Floor(src)
+        > ulong.MaxValue => throw new ArgumentOutOfRangeException(nameof(src), "Value is too large to be converted to ulong"),
+        < ulong.MinValue => throw new ArgumentOutOfRangeException(nameof(src), "Value is too small to be converted to ulong"),
+        _ => (ulong)Math.Floor(src)
     };
 
-    public static uint ToExpirationMilliseconds(this TimeToLive src) =>
+    public static ulong ToExpirationMilliseconds(this TimeToLive src) =>
         src.Value.TotalMilliseconds.ToExpirationMilliseconds();
 }